chore(refactor): move source types in source/types (refs #47)
